+++ chat2db-server/chat2db-spi/src/main/java/ai/chat2db/spi/model/TableMeta.java
@@ -19,4 +19,6 @@ public class TableMeta {
 
 
     private List<IndexType> indexTypes;
+
+    private List<DefaultValue> defaultValues;
 }